refactor: id 타입 Long으로 변경

+++ src/main/java/nextstep/web/controller/ReservationController.java
@@ -18,7 +18,7 @@ public class ReservationController {
 
     @PostMapping
     public ResponseEntity<URI> createReservation(@RequestBody ReservationRequestDto requestDto) {
-        int createdId = reservationService.createReservation(requestDto);
+        Long createdId = reservationService.createReservation(requestDto);
 
         return ResponseEntity.created(URI.create("/reservations/" + createdId))
                 .build();
